refactor: Rename package ftl.cli.firebase.test.android.configuration to ftl.cli.firebase.test.android.locales
