+++ test_runner/src/main/kotlin/ftl/cli/firebase/test/android/locales/AndroidLocalesDescribeCommand.kt
@@ -1,37 +1,41 @@
-package ftl.cli.firebase.test.android.configuration
+package ftl.cli.firebase.test.android.locales
 
 import ftl.config.FtlConstants
-import ftl.domain.ListAndroidLocales
+import ftl.domain.DescribeAndroidLocales
 import ftl.domain.invoke
 import picocli.CommandLine
 
 @CommandLine.Command(
-    name = "list",
+    name = "describe",
     headerHeading = "",
     synopsisHeading = "%n",
     descriptionHeading = "%n@|bold,underline Description:|@%n%n",
     parameterListHeading = "%n@|bold,underline Parameters:|@%n",
     optionListHeading = "%n@|bold,underline Options:|@%n",
-    header = ["Print current list of locales available to test against"],
-    description = ["Print current list of Android locales available to test against"],
+    header = ["Describe a locales "],
     usageHelpAutoWidth = true
 )
-class AndroidLocalesListCommand :
+class AndroidLocalesDescribeCommand :
     Runnable,
-    ListAndroidLocales {
+    DescribeAndroidLocales {
+
+    @CommandLine.Parameters(
+        index = "0",
+        arity = "1",
+        paramLabel = "LOCALE",
+        defaultValue = "",
+        description = [
+            "The locale to describe, found" +
+                " using \$ gcloud firebase test android locales list\n."
+        ]
+    )
+    override var locale: String = ""
 
     @CommandLine.Option(
         names = ["-c", "--config"],
         description = ["YAML config file path"]
     )
     override var configPath: String = FtlConstants.defaultAndroidConfig
-
-    @CommandLine.Option(
-        names = ["-h", "--help"],
-        usageHelp = true,
-        description = ["Prints this help message"]
-    )
-    var usageHelpRequested: Boolean = false
 
     override fun run() = invoke()
 }
